Rename content script to something more descriptive.
The reason of renaming it is because Google Chrome changed the
way we visualize the content script in the inspector. The inspector
removes the host portion of the URL leaving with just the filename.
That causes confusion since all my content scripts have the same
name. By making descriptive names, it would become easier to
locate the file within the inspector.