Fix bug when joining a hangout that was merged was incorrect.
The unique hangout identifier broke when I implemented the merge
feature for hangouts. The reason why it broke is because the
unique id depended on the length of the hangout array. Right now
it is changed to keep track of all available hangout notifications.
A hangout notification is anyone that reshares it. Merging wont
affect that.