Only refresh the data when the data has been received.
Searching takes time, it was resetting it first then searching
causing a slight delay when it is time to refresh the popup. This
only happened when the user keeps the popup opened or opens it
when the refresh is happening. This has changed now making it
only refresh when the data has arrived. Causing no flickering.