Cache address geo lookup to the extension db.
Before it was just caching the lat and long per browser session
so when the user restarts the browser, the cache is gone. Right
now it will persist in webStorage like how we do with other data
which will increase the memory consumption but will not request
same location if it is in storage.

Fixes #94